权限补充

+++ meet-model/src/main/java/com/nsu/stu/meet/model/vo/LimitVo.java
@@ -9,5 +9,5 @@ import lombok.*;
 public class LimitVo {
     private Long userId;
 
-    private Long limitId;
+    private Integer limitId;
 }
